Rename setup to flask.
This is so we can have more than one way to set this logger
and to different type of resources.

@rantonmattei
