+++ tests/test_logging.py
@@ -119,14 +119,6 @@ def test_configure_handler_without_dsn():
     assert isinstance(current_logging.handlers[0], logging.StreamHandler)
 
 
-def test_setting_up_flask_app():
-    """Test setting up a flask application.
-    """
-
-    app = Flask(__name__)
-    logger.setup(app, '', 'qa', 'logging', logging.INFO, 'service',  '1.0.0')
-
-
 def test_configure_handler_with_dsn():
     """Test configuration of the handler.
     """
@@ -135,40 +127,6 @@ def test_configure_handler_with_dsn():
     logger.configure_handler(current_logging, 'dsn', 'qa', 'logging', '1.0.0')
     assert isinstance(current_logging.handlers[0], logger.DSNHandler)
     assert current_logging.handlers[0].dsn == 'dsn'
-
-
-def test_app_correlation_id_creation():
-    """Test creation of the correlation id.
-    """
-
-    with app.test_request_context('/hello/'):
-        app.global_correlation_id()
-        assert g.correlation_id
-
-
-def test_app_correlation_id_reuse():
-    """Test correlation id reuse.
-
-    When passed from the header, the correlation id should be reused.
-    """
-
-    correlation_id = str(uuid.uuid1())
-    with app.test_request_context(
-            '/hello/', headers={'Correlation-Id': correlation_id}):
-        app.global_correlation_id()
-        assert g.correlation_id == correlation_id
-
-
-def test_logger_creation():
-    """Test creation of the logger.
-    """
-
-    with app.test_request_context('/hello/'):
-        app.global_logger()
-        assert g.correlation_id
-        assert isinstance(g.log, logging.LoggerAdapter)
-        assert 'correlation_id' in g.log.extra
-        assert g.log.extra.get('correlation_id') == g.correlation_id
 
 
 @pytest.mark.parametrize(('level', 'expected'), [
